Extract catalog booking runtime
